refactor(session): extract reference prompt helpers
